Using a context to pass values around the sheet editor.
This is to avoid the large amount of prop drilling that was amounting.
